pyproject.toml
[tool.poetry]
name = "flappy-bird-bot"
version = "0.1.0"
description = "Flappy Bird AI using DQN"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pygame = "^2.6.1"
numpy = "1.24.3"
tensorflow-gpu = "^2.10.0"
keras-rl2 = "^1.0.5"
protobuf = "3.19.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 